Feature/reward history

+++ src/main/java/com/ssg/starroadadmin/reward/entity/RewardHistory.java
@@ -22,5 +22,5 @@ public class RewardHistory extends BaseTimeEntity {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private int rewardId;
+    private Long rewardId;
 }
